requirements.txt
Django==4.2.7
mongoengine==0.27.0
pymongo==4.6.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5
django-environ==0.11.2
python-dotenv==1.0.0
djangorestframework-simplejwt==5.3.1
django-redis==5.4.0
celery==5.3.6
redis==5.0.1

stripe==7.11.0
requests==2.31.0
python-dateutil==2.8.2
django-cleanup==8.0.0
django-storages==1.14.2
django-celery-beat==2.5.0
django-celery-results==2.5.1
dnspython==2.6.1
gunicorn==21.2.0
whitenoise==6.6.0
python-jose==3.3.0
boto3==1.34.34